Commit v 28122020

+++ src/main/java/web/app/service/ReportService.java
@@ -81,7 +81,7 @@ public class ReportService {
             double fee = 0;
 
             try {
-                fee = dto.getFee().trim().isEmpty() ? 0:  nf.parse(dto.getFee()).doubleValue();
+                fee = dto.getFee().trim().isEmpty() ? 0: (double) nf.parse(dto.getFee()).doubleValue();
             }catch (Exception e)
             {
                 e.printStackTrace();
